+++ c-eyes@anaximeno/files/c-eyes@anaximeno/5.4/eyeModes.js
@@ -19,6 +19,8 @@
 
 const { Clutter } = imports.gi;
 
+const DEFAULT_VERTICAL_PADDING = 0.1;
+
 class EyeMode {
     /** @type Eye */
     eye;
@@ -72,9 +74,11 @@ class EyeMode {
 
 class EyelidMode extends EyeMode {
     drawEye(area, options) {
-        let [area_x, area_y] = this.areaPos();
+        let [mouse_x, mouse_y, mask] = global.get_pointer();
         let [area_width, area_height] = area.get_surface_size();
-        let [mouse_x, mouse_y, _] = global.get_pointer();
+        let [area_x, area_y] = this.areaPos();
+
+        const padding = options.padding || DEFAULT_VERTICAL_PADDING;
 
         area_x += area_width / 2;
         area_y += area_height / 2;
@@ -85,7 +89,7 @@ class EyelidMode extends EyeMode {
         let mouse_ang = Math.atan2(mouse_y, mouse_x);
         let mouse_rad = Math.sqrt(mouse_x * mouse_x + mouse_y * mouse_y);
 
-        let eye_rad = (area_height) / 2;
+        let eye_rad = (area_height - padding) / 2;
         let iris_rad = eye_rad * 0.5;
         let pupil_rad = iris_rad * 0.4;
 
@@ -166,9 +170,11 @@ class EyelidMode extends EyeMode {
 
 class EyelidFillMode extends EyeMode {
     drawEye(area, options) {
-        let [area_x, area_y] = this.areaPos();
+        let [mouse_x, mouse_y, mask] = global.get_pointer();
         let [area_width, area_height] = area.get_surface_size();
-        let [mouse_x, mouse_y, _] = global.get_pointer();
+        let [area_x, area_y] = this.areaPos();
+
+        const padding = options.padding || DEFAULT_VERTICAL_PADDING;
 
         area_x += area_width / 2;
         area_y += area_height / 2;
@@ -179,7 +185,7 @@ class EyelidFillMode extends EyeMode {
         let mouse_ang = Math.atan2(mouse_y, mouse_x);
         let mouse_rad = Math.sqrt(mouse_x * mouse_x + mouse_y * mouse_y);
 
-        let eye_rad = (area_height) / 2;
+        let eye_rad = (area_height - padding) / 2;
         let iris_rad = eye_rad * 0.5;
         let pupil_rad = iris_rad * 0.4;
 
@@ -263,13 +269,14 @@ class EyelidFillMode extends EyeMode {
 
 class BulbMode extends EyeMode {
     drawEye(area, options) {
-        let [area_x, area_y] = this.areaPos();
+        let [mouse_x, mouse_y, mask] = global.get_pointer();
         let [area_width, area_height] = area.get_surface_size();
+        let [area_x, area_y] = this.areaPos();
+
+        const padding = options.padding || DEFAULT_VERTICAL_PADDING;
 
         area_x += area_width / 2;
         area_y += area_height / 2;
-
-        let [mouse_x, mouse_y, _] = global.get_pointer();
 
         mouse_x -= area_x;
         mouse_y -= area_y;
@@ -277,7 +284,7 @@ class BulbMode extends EyeMode {
         let mouse_ang = Math.atan2(mouse_y, mouse_x);
         let mouse_rad = Math.sqrt(mouse_x * mouse_x + mouse_y * mouse_y);
 
-        let eye_rad = (area_height) / 2.3;
+        let eye_rad = (area_height - padding) / 2.3;
         let iris_rad = eye_rad * 0.6;
         let pupil_rad = iris_rad * 0.4;
 
@@ -333,13 +340,14 @@ class BulbMode extends EyeMode {
 
 class BulbFillMode extends EyeMode {
     drawEye(area, options) {
-        let [area_x, area_y] = this.areaPos();
+        let [mouse_x, mouse_y, mask] = global.get_pointer();
         let [area_width, area_height] = area.get_surface_size();
+        let [area_x, area_y] = this.areaPos();
+
+        const padding = options.padding || DEFAULT_VERTICAL_PADDING;
 
         area_x += area_width / 2;
         area_y += area_height / 2;
-
-        let [mouse_x, mouse_y, _] = global.get_pointer();
 
         mouse_x -= area_x;
         mouse_y -= area_y;
@@ -347,7 +355,7 @@ class BulbFillMode extends EyeMode {
         let mouse_ang = Math.atan2(mouse_y, mouse_x);
         let mouse_rad = Math.sqrt(mouse_x * mouse_x + mouse_y * mouse_y);
 
-        let eye_rad = (area_height) / 2.3;
+        let eye_rad = (area_height - padding) / 2.3;
         let iris_rad = eye_rad * 0.6;
         let pupil_rad = iris_rad * 0.4;