+++ multicore-sys-monitor@ccadeptic23/locale.sh
@@ -7,6 +7,7 @@ cd "po"
 intltool-extract --type=gettext/glade ../3.0/prefsui.glade
 intltool-extract --type=gettext/glade ../3.2/prefsui.glade
 intltool-extract --type=gettext/glade ../3.4/prefsui.glade
+intltool-extract --type=gettext/glade ../4.0/prefsui.glade
 xgettext --language=C --keyword=_ --keyword=N_ --output=multicore-sys-monitor.pot ../*/prefsui.glade.h
 xgettext --language=JavaScript --keyword=_ --keyword=N_ --output=multicore-sys-monitor.pot --join-existing --from-code=UTF-8 ../*/*.js ../metadata.json